requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
matplotlib>=3.4.0
tqdm>=4.60.0
tensorboard>=2.5.0
opencv-python>=4.5.0
pillow>=8.0.0
lpips>=0.1.4  # For perceptual loss
einops>=0.4.1  # For more readable tensor operations
imageio>=2.19.0  # For creating animations
imageio-ffmpeg>=0.4.7  # For creating MP4 videos